Separate InputField & TextField
